Added comments

+++ GNN/GraphCreation/pipeline/nx_to_torch.py
@@ -60,12 +60,12 @@ class Nx2T1Conv2: # this one also keeps track of node type
         if nx_graph.graph["value"] is not None:
              # TODO value should be int, but isnt allways
             y_val = int(round(nx_graph.graph["value"]))
-            nx_graph.graph["y"] = torch.tensor([y_val], dtype=torch.long)
+            nx_graph.graph["y"] = torch.tensor([y_val], dtype=torch.int)
             # print("y:",nx_graph.graph["value"], y_val)
         else:
             #TODO what do we do with None values?
             y_val = 3
-            nx_graph.graph["y"] = torch.tensor([y_val], dtype=torch.long)
+            nx_graph.graph["y"] = torch.tensor([y_val], dtype=torch.int)
 
 
         # x ~ features of nodes
